Implement auth repository
