+++ lib/domain/providers/service_locator.dart
@@ -3,7 +3,7 @@ import 'package:pro_tests/domain/providers/auth_state.dart';
 import 'package:pro_tests/domain/providers/test_attempt.dart';
 import 'package:pro_tests/domain/providers/test_creation.dart';
 import 'package:pro_tests/domain/providers/test_list.dart';
-import 'package:pro_tests/domain/repository/authentication.dart';
+import 'package:pro_tests/domain/repository/authentication/authentication.dart';
 import 'package:pro_tests/domain/services/auth_service.dart';
 
 abstract class ServiceLocator {
