+++ lib/ext.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart';
+
+extension ToTextWidget on String {
+  Widget toText() => Text(this);
+}
